+++ easygraphics/__init__.py
@@ -1,6 +1,6 @@
 __author__ = 'Roy Qu'
 __email__ = 'royqh1979@gmail.com'
-__version__ = '0.9.11'
+__version__ = '0.9.12'
 
 from .easygraphics import *